Chain of responsibility

+++ src/br/com/biancarosa/designpatterns/strategy/impostos/Imposto.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public interface Imposto {
 
-    BigDecimal calcula(Orcamento orcamento);
+    double calcula(Orcamento orcamento);
 
 }
